Add files via upload
Write a class, Grader, which has an instance variable, score, an appropriate constructor 
and appropriate methods. A method, letterGrade() that returns the letter grade as 
O/E/A/B/C/F.
Now write a demo class to test the Grader class by reading a score from the user, using 
it to create a Grader object after validating that the value is not negative and is not 
greater then 100. Finally, call the letterGrade() method to get and print the grade.